requirements.txt
nose==1.3.0
coverage==3.7
pylint==1.0.0
flake8==2.1.0
sphinx==1.2b3
mock==1.0.1
tornado==3.2.1
pyzmq==14.3.0
